[#35] 리뷰 반영
